requirements.txt
# Core Framework
flask>=2.3.0
flask-cors>=4.0.0

# OpenAI Client (for Ollama alternative)
openai>=1.3.0

# Environment and Configuration
python-dotenv>=1.0.0

# Data Processing and Storage
numpy>=1.24.0

# Web Server
werkzeug>=2.3.0
gunicorn>=21.2.0

# Document Processing and OCR
PyPDF2>=3.0.0
pdf2image>=1.16.0
pytesseract>=0.3.10
Pillow>=10.0.0
PyMuPDF>=1.23.0

# Text Processing and NLP
langdetect>=1.0.9
deep-translator>=1.11.4

# Language Detection (Alternative options)
langid>=1.1.6

# Text Splitting for RAG
langchain-text-splitters>=0.0.1

# HTTP Requests
requests>=2.31.0

# Email Support
email-validator>=2.0.0

# Password Security
bcrypt>=4.0.1

# Date and Time Utilities
python-dateutil>=2.8.2

# JSON Web Tokens (if needed for auth)
PyJWT>=2.8.0

# Database Migrations (optional)
alembic>=1.12.0

# Testing (optional)
pytest>=7.4.0
pytest-flask>=1.2.0

# Logging and Monitoring
structlog>=23.1.0

# Alternative AI Providers (choose one or more)
# Anthropic Claude
anthropic>=0.7.0

# Google AI
google-generativeai>=0.3.0

# Hugging Face Transformers
transformers>=4.35.0
torch>=2.1.0
sentence-transformers>=2.2.2

# Cohere
cohere>=4.37.0

# Together AI
together>=0.2.0

# Groq
groq>=0.4.0

# Azure OpenAI
# (removed: azure-ai-openai caused "No matching distribution found". 
#  Install manually or add a pinned compatible package/version if needed.)

# AWS Bedrock
boto3>=1.29.0

# Production WSGI Server
waitress>=2.1.2

# Security
cryptography>=41.0.0

# Performance
redis>=5.0.0
celery>=5.3.0

# Monitoring
sentry-sdk[flask]>=1.38.0

# Timezone Support
pytz>=2023.3

# Math and Statistics
scipy>=1.11.0

# HTTP Client with better error handling
httpx>=0.25.0

# Configuration Management
pydantic>=2.4.0
pydantic-settings>=2.0.0

# Async Support
aiohttp>=3.9.0

# Alternative Embeddings
chromadb>=0.4.0
faiss-cpu>=1.7.4

# Alternative to Ollama - Local LLM serving
ctransformers>=0.2.0

# OpenAI Alternative APIs
replicate>=0.15.0
